refactor: money as BigDecimal (ADR 0011, supersedes 0005)
Switch money from integer centavos to BigDecimal: clean field names
(monthlyRent/deposit/amount, no Cents suffix), stored as SQL NUMERIC, still
exact (never floating point). ADR 0005 superseded; all living docs updated;
Room.monthlyRent + Room DTOs changed to BigDecimal.

+++ src/main/kotlin/com/boardinghouse/backend/rooms/web/RoomResponse.kt
@@ -1,10 +1,12 @@
 package com.boardinghouse.backend.rooms.web
 
+import java.math.BigDecimal
+
 data class RoomResponse(
     val id: Long,
     val propertyId: Long,
     val label: String,
-    val monthlyRent: Long,
+    val monthlyRent: BigDecimal,
     val description: String,
     val active: Boolean,
 )
