+++ src/main/kotlin/com/boardinghouse/backend/rooms/web/UpdateRoomRequest.kt
@@ -1,7 +1,9 @@
 package com.boardinghouse.backend.rooms.web
 
+import java.math.BigDecimal
+
 data class UpdateRoomRequest(
     val label: String,
-    val monthlyRent: Long,
+    val monthlyRent: BigDecimal,
     val description: String,
 )
